requirements.txt
Django==1.8.0
MySQL-python==1.2.5
pylibmc==1.5.0
pyzmq==14.7.0
django-crispy-forms==1.4.0
addict==0.4.0
gunicorn==19.0.0
wsgiref==0.1.2
httplib2==0.9.1
simplejson==3.8.0
urllib3==1.12
whitenoise==2.0.3
pillow==2.9.0
djangorestframework==3.4.1
markdown==2.6.2
django-filter==0.11.0
django-rest-swagger==0.3.4
requests==2.7.0
python-dateutil==2.4.2
rsa==3.4.2
django-oauth-toolkit==0.10.0
docker-compose==1.9.0
pycrypto==2.6.1
fuzzyfinder==1.0.0
django-cors-headers==2.1.0
pypinyin==0.23.0













